Initial development version
This development version contains the first release of the core
framework that can be used for developing C# bots for Warlight's
AI challenge.

The framework has not yet been tested, and is therefore not ready
to be released to the master branch. Although several tasks are
open that should be finished before the framework can be used in
competitive environments.

For a list of those tasks, please see the project's home on GitHub:
https://github.com/jdno/AIChallengeFramework/issues