Fix bug while parsing pick_starting_regions
An exception was thrown in the method PickStartingRegions whenever
the game engine requested the list of preferred starting positions.
The cause for this behavior was that it was overlooked that the
engine actually provides the time the bot has to answer the request
as the first argument. As the time is currently 2000ms, this was
way out of bounds (# regions: 42).