Fix bug when parsing the opponent's moves
The parser automatically removes the first 15 characters from the
"opponent_moves ..." instruction. This is hepful, because the string
consists afterwards of only the moves. But if no moves by the
opponent are visible, the command has only 14 characters, which lead
to an exception. This has been fixed, and now 14 characters get
removed and then Trim is called on the string.