Rework check for the number of border regions
Continents initially didn't have a list of border regions, and
needed to calulcate the number of border regions on demand. With
the new way the two versions of the map are handled during the
game, and the list of border regions that is already available,
this complex approach is no longer needed, and is therefore
removed. Instead, simply the count of border regions is returned.