Fix bugs in Map
If a region is added that is already in the regions dictionary, an
exception is thrown. This can be prevented with a quick check if
the region's id is already a key in the dictionary.

And if you add an continent, its regions get added to the map's
dictionary of regions as well.